stm32: allow 16 bit SPI length requests

+++ src/stm32/spi.c
@@ -120,7 +120,7 @@ spi_prepare(struct spi_config config)
 
 void
 spi_transfer(struct spi_config config, uint8_t receive_data,
-             uint8_t len, uint8_t *data)
+             uint16_t len, uint8_t *data)
 {
     SPI_TypeDef *spi = config.spi;
     while (len--) {
